Use ptrace instead of ldpreload.
While trying to port tup to other platforms, I ran into some problems with some core binaries being statically linked. Not being able to support statically linked binaries was always an issue with the ldpreload method, but for example on FreeBSD even gcc is statically linked, which means using tup for a simple C program won't work.

Another issue is that I need a way for the child process to communicate the file accesses back to tup. I had been doing this with a socketpair, which worked well enough until someone tried to use a program that used glib, which closes all open file descriptors when it spawns. Another alternative here may have been just to write all file accesses to a file under .tup/, but with ptrace this issue doesn't come up. The reason is that the socket is used instead by the "parent" ptrace process (part of tup), rather than the child in the ldpreload library.

The other benefit to ptrace is that the list of calls that I need to intercept is well-defined, since I can just look up the list of syscalls. In contrast, trying to wrap stat() at the libc level could resolve to anything, since a header can just #define stat to be __xstat or whatever.

The main downside currently is speed. It's rather silly to have to call ptrace() to get the filename passed to sys_open() 4 or 8 bytes at a time, but that appears to be the way it's done.

Another downside is that the ptrace() interface is machine dependent. Apparently my program has to care which register is used to send the first arg to the syscall - I can't just get an array of them or whatever. Le sigh.