Buffer commands output
Currently when you run 'tup upd' jobs commands output logs directly to
stdout/stderr. This can be a problem if you run 'tup upd' with multiple
jobs - logs are mixed and it is harder to read what command produced
what log line.

This change redirects commands output to a temp file and later it dumps them
to stdout serially.