Buffer commands output
Currently when you run 'tup upd' jobs commands output logs directly to
stdout/stderr. This can be a problem if you run 'tup upd' with multiple
jobs - logs are mixed and it is harder to read what command produced
what log line.

This change redirects commands output to a temp file and later it dumps them
to stdout serially.

To make commands believe that their stdout/stderr is terminal we wrap isatty()
function with our own implementation that returns 'true' if the original
stdout/stderr is a terminal.